+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * [ADDED] support for defining default array item values in a schema definition
 * [ADDED] new `Instance.populate` method - handles loading of Model's associations
 * [ADDED] new `Model.getByIdOrFail` which returns rejected promise with `StorageError` when a document is not found in a bucket
+* [ADDED] new `Model.buildRefDocKey` method (Model's private method `$buildRefDocKey` has been updated and made public)
 
 * [ADDED] new `lean` option to `Model` "getByRefDoc" methods - if true, the method returns document's `Key` object instead of fetched document's data
 * [ADDED] new `refDocKey` Model option which allows us to provide custom constructor object used for generating reference document keys